+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changes
 =======
 
+Version 0.9.1 (UNRELEASED)
+--------------------------
+
+- Fixes conversion of possibly-negative resource usage values to human-readable formats.
+
 Version 0.9.0 (2022-12-13)
 --------------------------